requirements.txt
pandas==2.3.1
matplotlib==3.10.3
reportlab==4.4.3
gspread==6.0.2
oauth2client==4.1.3
pymysql==1.1.0
psycopg2-binary==2.9.9